supporting GEOS 3.10
FossilOrigin-Name: 19b042a1381f4d929a240583cbeb7fd4c0bb5d6b

+++ src/gaiageo/gg_wkt.c
@@ -73,7 +73,7 @@ gaiaOutClean (char *buffer)
  * integer numbers
 */
     int integer = 1;
-    for (i = 0; i < strlen (buffer); i++)
+    for (i = 0; i < (int)strlen (buffer); i++)
       {
 	  if (buffer[i] == '.')
 	      integer = 0;
